+++ probe-rs/src/architecture/riscv/mod.rs
@@ -539,6 +539,12 @@ impl<'probe> CoreInterface for Riscv32<'probe> {
 
         Ok(breakpoints)
     }
+
+    fn fpu_support(&mut self) -> Result<bool, crate::error::Error> {
+        Err(crate::error::Error::Other(anyhow::anyhow!(
+            "Fpu detection not yet implemented"
+        )))
+    }
 }
 
 impl<'probe> MemoryInterface for Riscv32<'probe> {