+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support added for trace funnels and SWO peripherals
   - Added custom sequencing for STM32H7 parts to configure debug system components on attach
 - Added support for ARMv8-A cores running in 64-bit mode (#1120)
+- Added FPU register reading support for cortex-m cores
 
 ### Changed